+++ Src/DependencyCollector/Shared/TelemetryDiagnosticSourceListener.cs
@@ -8,7 +8,7 @@ namespace Microsoft.ApplicationInsights.DependencyCollector.Implementation
     using Microsoft.ApplicationInsights.Extensibility;
     using Microsoft.ApplicationInsights.Extensibility.Implementation;
 
-    internal class TelemetryDiagnosticSourceListener : DiagnosticSourceListenerBase<HashSet<string>>
+    internal class TelemetryDiagnosticSourceListener : DiagnosticSourceListenerBase<HashSet<string>>, IDiagnosticEventHandler
     {
         internal const string ActivityStartNameSuffix = ".Start";
         internal const string ActivityStopNameSuffix = ".Stop";
@@ -19,6 +19,8 @@ namespace Microsoft.ApplicationInsights.DependencyCollector.Implementation
         private readonly Dictionary<string, HashSet<string>> includedDiagnosticSourceActivities 
             = new Dictionary<string, HashSet<string>>(StringComparer.OrdinalIgnoreCase);
 
+        private readonly Dictionary<string, IDiagnosticEventHandler> customEventHandlers = new Dictionary<string, IDiagnosticEventHandler>(StringComparer.OrdinalIgnoreCase);
+
         public TelemetryDiagnosticSourceListener(TelemetryConfiguration configuration, ICollection<string> includeDiagnosticSourceActivities) 
             : base(configuration)
         {
@@ -26,39 +28,19 @@ namespace Microsoft.ApplicationInsights.DependencyCollector.Implementation
             this.PrepareInclusionLists(includeDiagnosticSourceActivities);
         }
 
-        internal override bool IsSourceEnabled(DiagnosticListener value)
+        public bool IsEventEnabled(string evnt, object input1, object input2)
         {
-            return this.includedDiagnosticSources.Contains(value.Name);
+            return !evnt.EndsWith(ActivityStartNameSuffix, StringComparison.Ordinal);
         }
 
-        internal override bool IsEventEnabled(string evnt, object input1, object input2, DiagnosticListener diagnosticListener, HashSet<string> context)
+        public void OnEvent(KeyValuePair<string, object> evnt, DiagnosticListener diagnosticListener)
         {
-            return this.IsActivityIncluded(evnt, context)
-                && !evnt.EndsWith(ActivityStartNameSuffix, StringComparison.OrdinalIgnoreCase);
-        }
-
-        internal bool IsActivityIncluded(string activityName, HashSet<string> includedActivities)
-        {
-            // if no list of included activities then all are included
-            return includedActivities == null || includedActivities.Contains(activityName);
-        }
-
-        internal override void HandleEvent(KeyValuePair<string, object> evnt, DiagnosticListener diagnosticListener, HashSet<string> context)
-        {
-            if (!this.IsActivityIncluded(evnt.Key, context)
-                || !evnt.Key.EndsWith(ActivityStopNameSuffix, StringComparison.OrdinalIgnoreCase))
+            if (!evnt.Key.EndsWith(ActivityStopNameSuffix, StringComparison.Ordinal))
             {
                 return;
             }
 
             Activity currentActivity = Activity.Current;
-            if (currentActivity == null)
-            {
-                DependencyCollectorEventSource.Log.CurrentActivityIsNull();
-                return;
-            }
-
-            DependencyCollectorEventSource.Log.TelemetryDiagnosticSourceListenerActivityStopped(currentActivity.Id, currentActivity.OperationName);
 
             // extensibility point - can chain more telemetry extraction methods here
             ITelemetry telemetry = this.ExtractDependencyTelemetry(diagnosticListener, currentActivity);
@@ -76,6 +58,22 @@ namespace Microsoft.ApplicationInsights.DependencyCollector.Implementation
             telemetry.Context.Properties["Activity"] = currentActivity.OperationName;
 
             this.Client.Track(telemetry);
+        }
+
+        internal void RegisterHandler(string diagnosticSourceName, IDiagnosticEventHandler eventHandler)
+        {
+            this.customEventHandlers[diagnosticSourceName] = eventHandler;
+        }
+
+        internal override bool IsSourceEnabled(DiagnosticListener value)
+        {
+            return this.includedDiagnosticSources.Contains(value.Name);
+        }
+
+        internal override bool IsActivityEnabled(string activityName, HashSet<string> includedActivities)
+        {
+            // if no list of included activities then all are included
+            return includedActivities == null || includedActivities.Contains(activityName);
         }
 
         internal DependencyTelemetry ExtractDependencyTelemetry(DiagnosticListener diagnosticListener, Activity currentActivity)
@@ -207,6 +205,17 @@ namespace Microsoft.ApplicationInsights.DependencyCollector.Implementation
             }
 
             return includedActivities;
+        }
+
+        protected override IDiagnosticEventHandler GetEventHandler(string diagnosticListenerName)
+        {
+            IDiagnosticEventHandler eventHandler;
+            if (this.customEventHandlers.TryGetValue(diagnosticListenerName, out eventHandler))
+            {
+                return eventHandler;
+            }
+
+            return this;
         }
 
         private void PrepareInclusionLists(ICollection<string> includeDiagnosticSourceActivities)